refactor(automations): extract codex schedule/prompt form helpers to a tested lib
Moves parseCodexRrule, buildCodexRrule, splitAutomationPrompt, composeAutomationPrompt,
and RRULE_DAY_ORDER out of automations-view.tsx into a pure, tested lib at
src/lib/codex-automation-form.ts. Adds slugifyAutomationId. No behavior change.
